Restructure function hierarchy
Restructured the code so that getComputerChoice and getPlayerChoice
are now sub-functions of playRound in order for the playRound
function to be able to execute properly. Previously, it was not able
to reference the variables playerChoice and computerChoice as they were
defined in their respective functions so they were not able to be
referenced by playRound.